deploy decoder and create merkle root

+++ src/base/DecodersAndSanitizers/Protocols/ITB/aave/AaveDecoderAndSanitizer.sol
@@ -1,16 +1,14 @@
 /* SPDX-License-Identifier: UNLICENSED */
 pragma solidity ^0.8.0;
 
-import '../common/ITBContractDecoderAndSanitizer.sol';
+import "../common/ITBContractDecoderAndSanitizer.sol";
 
 abstract contract AaveDecoderAndSanitizer is ITBContractDecoderAndSanitizer {
-    function deposit(address asset, uint) external pure returns (bytes memory addressesFound)
-    {
+    function deposit(address asset, uint256) external pure returns (bytes memory addressesFound) {
         addressesFound = abi.encodePacked(asset);
     }
 
-    function withdrawSupply(address asset, uint) external pure returns (bytes memory addressesFound)
-    {
+    function withdrawSupply(address asset, uint256) external pure returns (bytes memory addressesFound) {
         addressesFound = abi.encodePacked(asset);
     }
 }
